+++ RELEASES.md
@@ -7,6 +7,9 @@ CFGU:
 AC:
 - `ssid`: Use `ACI_GetNetworkWirelessEssidSecuritySsid` instead of `ACU_GetSSID`
 
+Compiler:
+- Updated to be a little more smarter for fixing compilation (can only fix `error: cannot convert 'const std::nullopt_t' to` only.)
+
 # 1.2.0
 
 - Added this release markdown so i don't forget what i've done.